move handlers decoupled from environment to separate file
