+++ runtime/environment.go
@@ -21,8 +21,6 @@ package runtime
 import (
 	"time"
 
-	"go.opentelemetry.io/otel/attribute"
-
 	"github.com/onflow/cadence"
 	"github.com/onflow/cadence/activations"
 	"github.com/onflow/cadence/old_parser"
@@ -661,18 +659,6 @@ func (e *interpreterEnvironment) ResolveLocation(
 	return
 }
 
-func newCheckHandler(i *Interface) sema.CheckHandlerFunc {
-	return func(checker *sema.Checker, check func()) {
-		reportMetric(
-			check,
-			*i,
-			func(metrics Metrics, duration time.Duration) {
-				metrics.ProgramChecked(checker.Location, duration)
-			},
-		)
-	}
-}
-
 func (e *interpreterEnvironment) resolveImport(
 	_ *sema.Checker,
 	importedLocation common.Location,
@@ -855,19 +841,6 @@ func (e *interpreterEnvironment) newOnStatementHandler() interpreter.OnStatement
 	}
 }
 
-func newOnRecordTraceHandler(i *Interface) interpreter.OnRecordTraceFunc {
-	return func(
-		interpreter *interpreter.Interpreter,
-		functionName string,
-		duration time.Duration,
-		attrs []attribute.KeyValue,
-	) {
-		errors.WrapPanic(func() {
-			(*i).RecordTrace(functionName, interpreter.Location, duration, attrs)
-		})
-	}
-}
-
 func (e *interpreterEnvironment) NewAccountValue(
 	context interpreter.AccountCreationContext,
 	address interpreter.AddressValue,
@@ -962,18 +935,6 @@ func (e *interpreterEnvironment) newContractValueHandler() interpreter.ContractV
 	}
 }
 
-func newUUIDHandler(i *Interface) interpreter.UUIDHandlerFunc {
-	return func() (uuid uint64, err error) {
-		errors.WrapPanic(func() {
-			uuid, err = (*i).GenerateUUID()
-		})
-		if err != nil {
-			err = interpreter.WrappedExternalError(err)
-		}
-		return
-	}
-}
-
 func (e *interpreterEnvironment) newOnEventEmittedHandler() interpreter.OnEventEmittedFunc {
 	return func(
 		inter *interpreter.Interpreter,
@@ -988,45 +949,6 @@ func (e *interpreterEnvironment) newOnEventEmittedHandler() interpreter.OnEventE
 			eventValue,
 			e.runtimeInterface.EmitEvent,
 		)
-
-		return nil
-	}
-}
-
-func newInjectedCompositeFieldsHandler(accountHandler stdlib.AccountHandler) interpreter.InjectedCompositeFieldsHandlerFunc {
-	return func(
-		context interpreter.AccountCreationContext,
-		location Location,
-		_ string,
-		compositeKind common.CompositeKind,
-	) map[string]interpreter.Value {
-
-		switch compositeKind {
-		case common.CompositeKindContract:
-			var address Address
-
-			switch location := location.(type) {
-			case common.AddressLocation:
-				address = location.Address
-			default:
-				return nil
-			}
-
-			addressValue := interpreter.NewAddressValue(
-				context,
-				address,
-			)
-
-			return map[string]interpreter.Value{
-				sema.ContractAccountFieldName: stdlib.NewAccountReferenceValue(
-					context,
-					accountHandler,
-					addressValue,
-					interpreter.FullyEntitledAccountAccess,
-					interpreter.EmptyLocationRange,
-				),
-			}
-		}
 
 		return nil
 	}
@@ -1139,18 +1061,6 @@ func (e *interpreterEnvironment) newOnInvokedFunctionReturnHandler() func(_ *int
 	}
 }
 
-func newOnMeterComputation(i *Interface) interpreter.OnMeterComputationFunc {
-	return func(compKind common.ComputationKind, intensity uint) {
-		var err error
-		errors.WrapPanic(func() {
-			err = (*i).MeterComputation(compKind, intensity)
-		})
-		if err != nil {
-			panic(interpreter.WrappedExternalError(err))
-		}
-	}
-}
-
 func (e *interpreterEnvironment) InterpretContract(
 	location common.AddressLocation,
 	program *interpreter.Program,
@@ -1225,24 +1135,6 @@ func (e *interpreterEnvironment) newResourceOwnerChangedHandler() interpreter.On
 	}
 
 	return newResourceOwnerChangedHandler(&e.runtimeInterface)
-}
-
-func newResourceOwnerChangedHandler(i *Interface) interpreter.OnResourceOwnerChangeFunc {
-	return func(
-		interpreter *interpreter.Interpreter,
-		resource *interpreter.CompositeValue,
-		oldOwner common.Address,
-		newOwner common.Address,
-	) {
-		errors.WrapPanic(func() {
-			(*i).ResourceOwnerChanged(
-				interpreter,
-				resource,
-				oldOwner,
-				newOwner,
-			)
-		})
-	}
 }
 
 func (e *interpreterEnvironment) CommitStorage(inter *interpreter.Interpreter) error {
@@ -1327,123 +1219,4 @@ func (e *interpreterEnvironment) getBaseActivation(
 		baseActivation = e.defaultBaseActivation
 	}
 	return
-}
-
-func newCapabilityBorrowHandler(handler stdlib.CapabilityControllerHandler) interpreter.CapabilityBorrowHandlerFunc {
-
-	return func(
-		context interpreter.BorrowCapabilityControllerContext,
-		locationRange interpreter.LocationRange,
-		address interpreter.AddressValue,
-		capabilityID interpreter.UInt64Value,
-		wantedBorrowType *sema.ReferenceType,
-		capabilityBorrowType *sema.ReferenceType,
-	) interpreter.ReferenceValue {
-
-		return stdlib.BorrowCapabilityController(
-			context,
-			locationRange,
-			address,
-			capabilityID,
-			wantedBorrowType,
-			capabilityBorrowType,
-			handler,
-		)
-	}
-}
-
-func newCapabilityCheckHandler(handler stdlib.CapabilityControllerHandler) interpreter.CapabilityCheckHandlerFunc {
-	return func(
-		context interpreter.CheckCapabilityControllerContext,
-		locationRange interpreter.LocationRange,
-		address interpreter.AddressValue,
-		capabilityID interpreter.UInt64Value,
-		wantedBorrowType *sema.ReferenceType,
-		capabilityBorrowType *sema.ReferenceType,
-	) interpreter.BoolValue {
-
-		return stdlib.CheckCapabilityController(
-			context,
-			locationRange,
-			address,
-			capabilityID,
-			wantedBorrowType,
-			capabilityBorrowType,
-			handler,
-		)
-	}
-}
-
-func newValidateAccountCapabilitiesGetHandler(i *Interface) interpreter.ValidateAccountCapabilitiesGetHandlerFunc {
-	return func(
-		context interpreter.AccountCapabilityGetValidationContext,
-		locationRange interpreter.LocationRange,
-		address interpreter.AddressValue,
-		path interpreter.PathValue,
-		wantedBorrowType *sema.ReferenceType,
-		capabilityBorrowType *sema.ReferenceType,
-	) (bool, error) {
-		var (
-			ok  bool
-			err error
-		)
-		errors.WrapPanic(func() {
-			ok, err = (*i).ValidateAccountCapabilitiesGet(
-				context,
-				locationRange,
-				address,
-				path,
-				wantedBorrowType,
-				capabilityBorrowType,
-			)
-		})
-		if err != nil {
-			err = interpreter.WrappedExternalError(err)
-		}
-		return ok, err
-	}
-}
-
-func newValidateAccountCapabilitiesPublishHandler(i *Interface) interpreter.ValidateAccountCapabilitiesPublishHandlerFunc {
-	return func(
-		context interpreter.AccountCapabilityPublishValidationContext,
-		locationRange interpreter.LocationRange,
-		address interpreter.AddressValue,
-		path interpreter.PathValue,
-		capabilityBorrowType *interpreter.ReferenceStaticType,
-	) (bool, error) {
-		var (
-			ok  bool
-			err error
-		)
-		errors.WrapPanic(func() {
-			ok, err = (*i).ValidateAccountCapabilitiesPublish(
-				context,
-				locationRange,
-				address,
-				path,
-				capabilityBorrowType,
-			)
-		})
-		if err != nil {
-			err = interpreter.WrappedExternalError(err)
-		}
-		return ok, err
-	}
-}
-
-func configureVersionedFeatures(i Interface) {
-	var (
-		minimumRequiredVersion string
-		err                    error
-	)
-	errors.WrapPanic(func() {
-		minimumRequiredVersion, err = i.MinimumRequiredVersion()
-	})
-	if err != nil {
-		panic(err)
-	}
-
-	// No feature flags yet
-	_ = minimumRequiredVersion
 }
